Parametrizing tests. Moving test module to resources.
